Update new test to use precheck functions Move `TestBackendEncryptionKeyEmptyConflict` test to internal test file
